requirements_dev.txt
coverage==7.6.4
cupy-cuda12x==13.3.0
flake8==7.1.1
mypy==1.12.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-html==4.1.1
tox==4.23.1
